Expose table reading functionality
